Move thread input requeue behind inlet
