+++ backend/threads/chat_adapters/runtime_thread_input_action.py
@@ -42,20 +42,6 @@ def dispatch_queued_thread_input_action(queue_manager: Any, action: QueuedThread
     return {"status": "queued", "thread_id": action.thread_id}
 
 
-def requeue_thread_input_item(queue_manager: Any, thread_id: str, item: Any) -> None:
-    queue_manager.enqueue(
-        item.content,
-        thread_id,
-        notification_type=item.notification_type,
-        source=item.source,
-        sender_id=item.sender_id,
-        sender_name=item.sender_name,
-        sender_avatar_url=item.sender_avatar_url,
-        is_steer=item.is_steer,
-        metadata=item.metadata,
-    )
-
-
 def owner_runtime_thread_input_action(
     *,
     thread_id: str,
