+++ backend/threads/run/cancellation.py
@@ -2,7 +2,7 @@ import asyncio
 import json
 from typing import Any
 
-from backend.threads.chat_adapters.runtime_thread_input_action import requeue_thread_input_item
+from backend.threads.chat_adapters.thread_input_inlet import requeue_thread_input_item
 from core.runtime.notifications import is_terminal_background_notification
 
 
@@ -115,7 +115,7 @@ async def flush_cancelled_owner_steers(
     await persist_cancelled_owner_steers(agent=agent, config=config, items=owner_steers)
 
     for item in passthrough:
-        requeue_thread_input_item(qm, thread_id, item)
+        requeue_thread_input_item(qm, thread_id=thread_id, item=item)
 
 
 async def emit_queued_terminal_followups(
@@ -130,7 +130,7 @@ async def emit_queued_terminal_followups(
         queued_items = app.state.queue_manager.drain_all(thread_id)
         extra_terminal, passthrough = partition_terminal_followups(queued_items)
         for item in passthrough:
-            requeue_thread_input_item(app.state.queue_manager, thread_id, item)
+            requeue_thread_input_item(app.state.queue_manager, thread_id=thread_id, item=item)
         for item in extra_terminal:
             await emit(
                 {
